+++ ChromaRendererCore/src/ISpacePartitioningStructure.h
@@ -15,7 +15,7 @@ class ISpacePartitioningStructure : public IIntersectable
     // virtual bool build(std::vector<Object>& objects) = 0;
     virtual bool build(std::vector<Mesh*>& meshes) = 0;
     // virtual bool intersect(Ray& r, Intersection& intersection) const = 0;
-    virtual long sizeInBytes(void) = 0;
+    virtual size_t sizeInBytes(void) = 0;
     // virtual void abort(void) = 0;
     virtual ~ISpacePartitioningStructure(){};
 };
